Support CPU inference with VSX PowerPC ISA (#5652)
